+++ src/main/java/com/example/restaurantVotingApplicationOnSpringBoot/to/menu/AdminMenuDto.java
@@ -15,7 +15,7 @@ public class AdminMenuDto extends AdminMenuDtoWithoutRestaurantId {
     @NotNull
     private Integer restaurantId;
 
-    public AdminMenuDto(Integer id, String name, Boolean enabled, Integer restaurantId, Double price) {
+    public AdminMenuDto(Integer id, String name, Boolean enabled, Integer restaurantId, Long price) {
         super(id, name, enabled, price);
         this.restaurantId = restaurantId;
         this.price = price;
